requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
requests==2.31.0
slowapi==0.1.9
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.0
mangum>=0.17.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.1
email-validator==2.1.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
aiofiles==23.2.1
websockets==12.0
